Uniswap addLiquidity integrated
Finally got the uniswap contracts mocked (gave up on mocking the whole infrastructure, it was on like 5 different sol versions. Made a mock that did the basics and moved on. 
Next up is figuring out why its not calling the transition contract correctly (see console logs in tests, wtf is not emited when it very much should be).